requirements.txt
streamlit>=1.37.0
pydantic>=2.8.0
openai>=1.40.0
python-dotenv>=1.0.1
plotly>=5.22.0
pandas>=2.2.2
httpx==0.27.2
requests>=2.32.3
beautifulsoup4>=4.12.3